requirements.txt
Pillow==8.0.1
pytest==6.1.2
pytest-subtests==0.3.2
